Cleaned up admin controller
